INT-3088 (S)FTP Outbound Gateway - PUT and MPUT
- Core support in file module
- FTP Parser and Test

https://jira.springsource.org/browse/INT-3088

+++ spring-integration-ftp/src/test/java/org/springframework/integration/ftp/TesFtpServer.java
@@ -91,7 +91,7 @@ public class TesFtpServer {
 				file.createNewFile();
 
 				targetFtpDirectory = new File(ftpRootFolder, "ftpTarget");
-				targetFtpDirectory.mkdirs();
+				targetFtpDirectory.mkdir();
 			}
 		};
 		this.localFolder  = new TemporaryFolder() {
@@ -113,7 +113,7 @@ public class TesFtpServer {
 				file.createNewFile();
 
 				targetLocalDirectory = new File(rootFolder, "localTarget");
-				targetLocalDirectory.mkdirs();
+				targetLocalDirectory.mkdir();
 			}
 		};
 	}
@@ -173,14 +173,16 @@ public class TesFtpServer {
 	}
 
 
-	public static void recursiveDelete(File file) {
+	public void recursiveDelete(File file) {
 		File[] files = file.listFiles();
 		if (files != null) {
 			for (File each : files) {
 				recursiveDelete(each);
 			}
 		}
-		file.delete();
+		if (!(file.equals(this.targetFtpDirectory) || file.equals(this.targetLocalDirectory))) {
+			file.delete();
+		}
 	}
